Refactor auth packages
