+++ src/main/java/com/api/prisma_vi/auth/AuthenticationService.java
@@ -1,7 +1,9 @@
 package com.api.prisma_vi.auth;
 
+import com.api.prisma_vi.auth.register.RegisterForm;
 import com.api.prisma_vi.user.UsersRepository;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.ResponseEntity;
 import org.springframework.security.core.userdetails.UserDetails;
 import org.springframework.security.core.userdetails.UserDetailsService;
 import org.springframework.security.core.userdetails.UsernameNotFoundException;
@@ -17,4 +19,5 @@ public class AuthenticationService implements UserDetailsService {
     public UserDetails loadUserByUsername(String username) throws UsernameNotFoundException {
         return repository.findByEmail(username);
     }
+
 }
