+++ src/main/java/com/api/prisma_vi/auth/login/AuthenticationView.kt
@@ -1,3 +1,3 @@
-package com.api.prisma_vi.auth
+package com.api.prisma_vi.auth.login
 
 data class AuthenticationView(val token: String, val prefix: String)
