+++ src/main/java/com/api/prisma_vi/auth/login/AutheticationForm.kt
@@ -1,4 +1,4 @@
-package com.api.prisma_vi.auth
+package com.api.prisma_vi.auth.login
 
 @JvmRecord
  data class AutheticationForm(
